move FiltersV2 to rule-list directory
