+++ public/app/features/alerting/unified/components/rules/Filter/RulesFilter.tsx
@@ -5,7 +5,7 @@ import { config } from '@grafana/runtime';
 import RulesFilterV1 from './RulesFilter.v1';
 import { SupportedView } from './RulesViewModeSelector';
 
-const RulesFilterV2 = lazy(() => import('./RulesFilter.v2'));
+const RulesFilterV2 = lazy(() => import('../../../rule-list/filter/RulesFilter.v2'));
 
 export interface RulesFilterProps {
   onClear?: () => void;
